moved deblur and added eval tools
