+++ deblur/deblur_cpu/.gitignore
@@ -1,0 +1,2 @@
+*.jpg
+*.png
