+++ deblur/deblur_cpu/main.h
@@ -1,0 +1,6 @@
+#ifndef MAIN_H
+#define MAIN_H
+
+typedef unsigned char byte;
+
+#endif // MAIN_H
