+++ deblur/deblur_cuda/.gitignore
@@ -1,0 +1,2 @@
+deblur
+*.png
