--- deblur_cpu/.gitignore
@@ -1,2 +1,0 @@
-*.jpg
-*.png
