--- deblur_cpu/main.h
@@ -1,6 +1,0 @@
-#ifndef MAIN_H
-#define MAIN_H
-
-typedef unsigned char byte;
-
-#endif // MAIN_H
