--- deblur_cuda/.gitignore
@@ -1,2 +1,0 @@
-deblur
-*.png
