--- deblur_cuda/matlab_test/.gitignore
@@ -1,3 +1,0 @@
-*.asv
-*.jpg
-*.png
